aulas/Aulas/DL - Agentes/DL - Rag/requirements.txt
streamlit 
llama-index-core 
llama-index-llms-ollama 
llama-index-embeddings-ollama
llama-index-readers-file 
pymupdf
langchain
langchain-community
langchain-ollama
faiss-cpu